+++ spectator-reg-atlas/src/main/java/com/netflix/spectator/atlas/AtlasRegistry.java
@@ -28,6 +28,11 @@ import com.netflix.spectator.api.Id;
 import com.netflix.spectator.api.Measurement;
 import com.netflix.spectator.api.Tag;
 import com.netflix.spectator.api.Timer;
+import com.netflix.spectator.atlas.impl.EvalPayload;
+import com.netflix.spectator.atlas.impl.MeasurementSerializer;
+import com.netflix.spectator.atlas.impl.PublishPayload;
+import com.netflix.spectator.atlas.impl.Subscription;
+import com.netflix.spectator.atlas.impl.Subscriptions;
 import com.netflix.spectator.impl.AsciiSet;
 import com.netflix.spectator.impl.Scheduler;
 import com.netflix.spectator.sandbox.HttpClient;
