+++ spectator-reg-atlas/src/test/java/com/netflix/spectator/atlas/MeasurementSerializerTest.java
@@ -19,8 +19,9 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.module.SimpleModule;
 import com.netflix.spectator.api.DefaultRegistry;
 import com.netflix.spectator.api.Id;
-import com.netflix.spectator.api.ManualClock;
 import com.netflix.spectator.api.Measurement;
+import com.netflix.spectator.atlas.impl.MeasurementSerializer;
+import com.netflix.spectator.atlas.impl.PublishPayload;
 import com.netflix.spectator.impl.AsciiSet;
 import org.junit.Assert;
 import org.junit.Test;
